更新

+++ src/main/java/com/shop/dao/ProductDao.java
@@ -10,7 +10,7 @@ public class ProductDao extends DBUtil {
 	/**
 	 * 插入一个product对象
 	 * 
-	 * @param cust
+	 * @param
 	 * @return
 	 */
 	public boolean insert(Product p) {
@@ -126,8 +126,8 @@ public class ProductDao extends DBUtil {
 				p.setpKeepTime(rs.getInt(6));
 				p.setpImage(rs.getString(7));
 				p.setkId(rs.getInt(8));
-				p.setiPrice(rs.getBigDecimal(9));
-				p.setmPrice(rs.getBigDecimal(10));
+				p.setiPrice(rs.getDouble(9));
+				p.setmPrice(rs.getDouble(10));
 				p.setIsHot(rs.getString(11));
 				p.setIsShow(rs.getString(12));
 				return p;
@@ -143,7 +143,7 @@ public class ProductDao extends DBUtil {
 	/**
 	 * 按照Name查找商品
 	 * 
-	 * @param name 商品名
+	 * @param
 	 * @return
 	 */
 	public Product find(String pname) {
@@ -161,8 +161,8 @@ public class ProductDao extends DBUtil {
 				p.setpKeepTime(rs.getInt(6));
 				p.setpImage(rs.getString(7));
 				p.setkId(rs.getInt(8));
-				p.setiPrice(rs.getBigDecimal(9));
-				p.setmPrice(rs.getBigDecimal(10));
+				p.setiPrice(rs.getDouble(9));
+				p.setmPrice(rs.getDouble(10));
 				p.setIsHot(rs.getString(11));
 				p.setIsShow(rs.getString(12));
 				return p;
@@ -195,8 +195,8 @@ public class ProductDao extends DBUtil {
 				p.setpKeepTime(rs.getInt(6));
 				p.setpImage(rs.getString(7));
 				p.setkId(rs.getInt(8));
-				p.setiPrice(rs.getBigDecimal(9));
-				p.setmPrice(rs.getBigDecimal(10));
+				p.setiPrice(rs.getDouble(9));
+				p.setmPrice(rs.getDouble(10));
 				p.setIsHot(rs.getString(11));
 				p.setIsShow(rs.getString(12));
 				list.add(p);
@@ -230,8 +230,8 @@ public class ProductDao extends DBUtil {
 				p.setpKeepTime(rs.getInt(6));
 				p.setpImage(rs.getString(7));
 				p.setkId(rs.getInt(8));
-				p.setiPrice(rs.getBigDecimal(9));
-				p.setmPrice(rs.getBigDecimal(10));
+				p.setiPrice(rs.getDouble(9));
+				p.setmPrice(rs.getDouble(10));
 				p.setIsHot(rs.getString(11));
 				p.setIsShow(rs.getString(12));
 				list.add(p);
@@ -265,8 +265,8 @@ public class ProductDao extends DBUtil {
 				p.setpKeepTime(rs.getInt(6));
 				p.setpImage(rs.getString(7));
 				p.setkId(rs.getInt(8));
-				p.setiPrice(rs.getBigDecimal(9));
-				p.setmPrice(rs.getBigDecimal(10));
+				p.setiPrice(rs.getDouble(9));
+				p.setmPrice(rs.getDouble(10));
 				p.setIsHot(rs.getString(11));
 				p.setIsShow(rs.getString(12));
 				list.add(p);
@@ -314,8 +314,8 @@ public class ProductDao extends DBUtil {
 			p.setpKeepTime(rs.getInt(6));
 			p.setpImage(rs.getString(7));
 			p.setkId(rs.getInt(8));
-			p.setiPrice(rs.getBigDecimal(9));
-			p.setmPrice(rs.getBigDecimal(10));
+			p.setiPrice(rs.getDouble(9));
+			p.setmPrice(rs.getDouble(10));
 			p.setIsHot(rs.getString(11));
 			p.setIsShow(rs.getString(12));
 			return p;
